+++ src/main/java/de/mossgrabers/convertwithmoss/core/creator/DestinationAudioFormat.java
@@ -1,0 +1,186 @@
+// Written by Jürgen Moßgraber - mossgrabers.de
+// (c) 2019-2024
+// Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
+
+package de.mossgrabers.convertwithmoss.core.creator;
+
+/**
+ * Settings for destination audio files.
+ *
+ * @author Jürgen Moßgraber
+ */
+public class DestinationAudioFormat
+{
+    private boolean updateBroadcastAudioChunk;
+    private boolean updateInstrumentChunk;
+    private boolean updateSampleChunk;
+    private boolean removeJunkChunks;
+
+    private int []  bitResolutions = null;
+    private int     maxSampleRate  = -1;
+    private boolean upSample;
+
+
+    /**
+     * Default constructor.
+     */
+    public DestinationAudioFormat ()
+    {
+        // Intentionally empty
+    }
+
+
+    /**
+     * Constructor.
+     *
+     * @param updateBroadcastAudioChunk Add or update the broadcast audio extension chunk if true
+     * @param updateInstrumentChunk Add or update the instrument chunk if true
+     * @param updateSampleChunk If true the sample chunk is add or updated if already present
+     * @param removeJunkChunks If true remove JUNK, junk, FLLR and MD5 chunks
+     */
+    public DestinationAudioFormat (final boolean updateBroadcastAudioChunk, final boolean updateInstrumentChunk, final boolean updateSampleChunk, final boolean removeJunkChunks)
+    {
+        this.updateBroadcastAudioChunk = updateBroadcastAudioChunk;
+        this.updateInstrumentChunk = updateInstrumentChunk;
+        this.updateSampleChunk = updateSampleChunk;
+        this.removeJunkChunks = removeJunkChunks;
+    }
+
+
+    /**
+     * Constructor.
+     *
+     * @param bitResolutions The allowed destination resolutions
+     * @param maxSampleRate The maximum supported sample rate
+     * @param upSample If true, the wave form is up-sampled if it is lower than the maximum sample
+     *            rate
+     */
+    public DestinationAudioFormat (final int [] bitResolutions, final int maxSampleRate, final boolean upSample)
+    {
+        this.bitResolutions = bitResolutions;
+        this.maxSampleRate = maxSampleRate;
+        this.upSample = upSample;
+    }
+
+
+    /**
+     * Should the broadcast audio chunk be updated
+     *
+     * @return True if the broadcast audio chunk be updated
+     */
+    public boolean isUpdateBroadcastAudioChunk ()
+    {
+        return this.updateBroadcastAudioChunk;
+    }
+
+
+    /**
+     * Should the instrument chunk be updated
+     *
+     * @return True if the instrument chunk be updated
+     */
+    public boolean isUpdateInstrumentChunk ()
+    {
+        return this.updateInstrumentChunk;
+    }
+
+
+    /**
+     * Should the sample chunk be updated
+     *
+     * @return True if the sample chunk be updated
+     */
+    public boolean isUpdateSampleChunk ()
+    {
+        return this.updateSampleChunk;
+    }
+
+
+    /**
+     * Shall junk chunks be removed?
+     *
+     * @return True if junk chunks should be removed
+     */
+    public boolean isRemoveJunkChunks ()
+    {
+        return this.removeJunkChunks;
+    }
+
+
+    /**
+     * Check if either a chunk option is enabled or bit resolution / frequency needs to be
+     * re-sampled.
+     *
+     * @return True if at least one is enabled
+     */
+    public boolean requiresRewrite ()
+    {
+        return this.updateBroadcastAudioChunk || this.updateInstrumentChunk || this.updateSampleChunk || this.removeJunkChunks || this.bitResolutions != null || this.maxSampleRate != -1;
+    }
+
+
+    /**
+     * Get the allowed destination resolutions.
+     *
+     * @return The resolutions as bits (e.g. 24 bit)
+     */
+    public int [] getBitResolutions ()
+    {
+        return this.bitResolutions;
+    }
+
+
+    /**
+     * Set the allowed destination resolutions.
+     *
+     * @param bitResolutions The resolutions to set
+     */
+    public void setBitResolutions (final int [] bitResolutions)
+    {
+        this.bitResolutions = bitResolutions;
+    }
+
+
+    /**
+     * Get the maximum supported sample rate.
+     *
+     * @return The maximum sample rate
+     */
+    public int getMaxSampleRate ()
+    {
+        return this.maxSampleRate;
+    }
+
+
+    /**
+     * Set the maximum supported sample rate.
+     *
+     * @param maxSampleRate The maximum sample rate to set
+     */
+    public void setMaxSampleRate (final int maxSampleRate)
+    {
+        this.maxSampleRate = maxSampleRate;
+    }
+
+
+    /**
+     * If true, the wave form is up-sampled if it is lower than the maximum sample rate.
+     *
+     * @return True to up-sample
+     */
+    public boolean isUpSample ()
+    {
+        return this.upSample;
+    }
+
+
+    /**
+     * Set the up-sample flag.
+     *
+     * @param upSample If true, up-sampling is enabled
+     */
+    public void setUpSample (final boolean upSample)
+    {
+        this.upSample = upSample;
+    }
+}