+++ src/main/java/de/mossgrabers/convertwithmoss/format/sfz/SfzCreator.java
@@ -19,6 +19,16 @@ import de.mossgrabers.convertwithmoss.core.model.enumeration.FilterType;
 import de.mossgrabers.convertwithmoss.core.model.enumeration.LoopType;
 import de.mossgrabers.convertwithmoss.core.model.enumeration.PlayLogic;
 import de.mossgrabers.convertwithmoss.core.model.enumeration.TriggerType;
+import de.mossgrabers.tools.ui.BasicConfig;
+import de.mossgrabers.tools.ui.control.TitledSeparator;
+import de.mossgrabers.tools.ui.panel.BoxPanel;
+
+import javafx.geometry.Orientation;
+import javafx.scene.Node;
+import javafx.scene.control.CheckBox;
+
+import javax.sound.sampled.AudioFileFormat;
+import javax.sound.sampled.UnsupportedAudioFileException;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -39,16 +49,18 @@ import java.util.Optional;
  */
 public class SfzCreator extends AbstractCreator
 {
-    private static final char                    LINE_FEED       = '\n';
-    private static final String                  FOLDER_POSTFIX  = " Samples";
-    private static final String                  SFZ_HEADER      = """
+    private static final String                  SFZ_CONVERT_TO_FLAC = "SfzConvertToFlac";
+    private static final AudioFileFormat.Type    TARGET_FORMAT       = new AudioFileFormat.Type ("FLAC", "flac");
+    private static final char                    LINE_FEED           = '\n';
+    private static final String                  FOLDER_POSTFIX      = " Samples";
+    private static final String                  SFZ_HEADER          = """
             /////////////////////////////////////////////////////////////////////////////
             ////
             """;
-    private static final String                  COMMENT_PREFIX  = "//// ";
+    private static final String                  COMMENT_PREFIX      = "//// ";
 
-    private static final Map<FilterType, String> FILTER_TYPE_MAP = new EnumMap<> (FilterType.class);
-    private static final Map<LoopType, String>   LOOP_TYPE_MAP   = new EnumMap<> (LoopType.class);
+    private static final Map<FilterType, String> FILTER_TYPE_MAP     = new EnumMap<> (FilterType.class);
+    private static final Map<LoopType, String>   LOOP_TYPE_MAP       = new EnumMap<> (LoopType.class);
 
     static
     {
@@ -62,6 +74,8 @@ public class SfzCreator extends AbstractCreator
         LOOP_TYPE_MAP.put (LoopType.ALTERNATING, "alternate");
     }
 
+    private CheckBox convertToFlac;
+
 
     /**
      * Constructor.
@@ -71,6 +85,41 @@ public class SfzCreator extends AbstractCreator
     public SfzCreator (final INotifier notifier)
     {
         super ("SFZ", notifier);
+    }
+
+
+    /** {@inheritDoc} */
+    @Override
+    public Node getEditPane ()
+    {
+        final BoxPanel panel = new BoxPanel (Orientation.VERTICAL);
+
+        this.convertToFlac = panel.createCheckBox ("@IDS_SFZ_CONVERT_TO_FLAC");
+
+        final TitledSeparator separator = this.addWavChunkOptions (panel);
+        separator.getStyleClass ().add ("titled-separator-pane");
+
+        return panel.getPane ();
+    }
+
+
+    /** {@inheritDoc} */
+    @Override
+    public void loadSettings (final BasicConfig config)
+    {
+        this.convertToFlac.setSelected (config.getBoolean (SFZ_CONVERT_TO_FLAC, false));
+
+        this.loadWavChunkSettings (config, "Sfz");
+    }
+
+
+    /** {@inheritDoc} */
+    @Override
+    public void saveSettings (final BasicConfig config)
+    {
+        config.setBoolean (SFZ_CONVERT_TO_FLAC, this.convertToFlac.isSelected ());
+
+        this.saveWavChunkSettings (config, "Sfz");
     }
 
 
@@ -99,7 +148,20 @@ public class SfzCreator extends AbstractCreator
         // Store all samples
         final File sampleFolder = new File (destinationFolder, safeSampleFolderName);
         safeCreateDirectory (sampleFolder);
-        this.writeSamples (sampleFolder, multisampleSource, true, false, false, false);
+
+        if (this.convertToFlac.isSelected ())
+        {
+            try
+            {
+                this.writeSamples (sampleFolder, multisampleSource, TARGET_FORMAT);
+            }
+            catch (final UnsupportedAudioFileException ex)
+            {
+                throw new IOException (ex);
+            }
+        }
+        else
+            this.writeSamples (sampleFolder, multisampleSource, this.getChunkSettings ());
 
         this.notifier.log ("IDS_NOTIFY_PROGRESS_DONE");
     }
@@ -185,8 +247,10 @@ public class SfzCreator extends AbstractCreator
      */
     private void createSample (final String safeSampleFolderName, final StringBuilder buffer, final ISampleZone zone, final int sequenceNumber)
     {
+        final String ending = this.convertToFlac.isSelected () ? ".flac" : ".wav";
+
         buffer.append ("\n<").append (SfzHeader.REGION).append (">\n");
-        addAttribute (buffer, SfzOpcode.SAMPLE, AbstractCreator.formatFileName (safeSampleFolderName, zone.getName () + ".wav"), true);
+        addAttribute (buffer, SfzOpcode.SAMPLE, AbstractCreator.formatFileName (safeSampleFolderName, zone.getName () + ending), true);
 
         // Default is 'attack' and does not need to be added
         final TriggerType trigger = zone.getTrigger ();